// swiyu-ecosystem-portal-ui/src/app/api/generated/model/trust-onboarding-submission.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Address} from './address';
import {Contact} from './contact';
import {MultiLanguageText} from './multi-language-text';
import {ProofOfPossession} from './proof-of-possession';

export interface TrustOnboardingSubmission {
  id: string;
  version: number;
  partnerId: string;
  entityName: MultiLanguageText;
  entityEmail: string;
  entityAddress: Address;
  contactPerson: Contact;
  status: TrustOnboardingSubmission.StatusEnum;
  proofOfPossessionList: Array<ProofOfPossession>;
  isGovActor: boolean;
  registryIds: {[key: string]: string};
  rejectionReason?: string;
  declineReason?: string;
  partnerNote?: string;
  correspondingLanguage?: TrustOnboardingSubmission.CorrespondingLanguageEnum;
  initiatedAt?: string;
  submittedAt?: string;
}
export namespace TrustOnboardingSubmission {
  export const StatusEnum = {
    Unsubmitted: 'UNSUBMITTED',
    Submitted: 'SUBMITTED',
    Succeeded: 'SUCCEEDED',
    Rejected: 'REJECTED',
    InformationRequested: 'INFORMATION_REQUESTED'
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
  export const CorrespondingLanguageEnum = {
    En: 'EN',
    De: 'DE',
    Fr: 'FR',
    It: 'IT',
    Rm: 'RM'
  } as const;
  export type CorrespondingLanguageEnum = (typeof CorrespondingLanguageEnum)[keyof typeof CorrespondingLanguageEnum];
}